requirements.txt
Flask
nltk
libmagic
python-magic
